requirements.txt
# requirements.txt

aerich==0.7.2
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
atlastk==0.13.3
attrs==23.2.0
bidict==0.23.1
bottle==0.12.25
captcha==0.5.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
cryptography==42.0.5
dictdiffer==0.9.0
docutils==0.19
fastapi==0.109.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
ifaddr==0.2.0
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
libsass==0.23.0
markdown2==2.4.13
MarkupSafe==2.1.5
multidict==6.0.5
nicegui==1.4.23
numpy==1.26.4
orjson==3.9.15
pandas==2.2.1
pillow==10.3.0
proxy_tools==0.1.0
pscript==0.7.7
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg-pool==3.2.1
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pypika-tortoise==0.1.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-multipart==0.0.9
python-socketio==5.11.1
pythonnet==3.0.3
pytz==2024.1
pywebview==5.0.5
PyYAML==6.0.1
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
tomlkit==0.12.4
tortoise-orm==0.20.0
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.29.0
vbuild==0.8.2
watchfiles==0.21.0
websockets==12.0
wsproto==1.2.0
yarl==1.9.4
